Fix: Rename file (#36)
* Rename file

* Update yarn.lock
